又是更新一次 更新了更新功能

+++ app/src/main/java/homes/gensokyo/enigma/viewmodel/UsrdataModel.kt
@@ -7,12 +7,14 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
+import homes.gensokyo.enigma.BuildConfig
 import homes.gensokyo.enigma.MainApplication.Companion.context
 import homes.gensokyo.enigma.logic.logic.UserRepository
 import homes.gensokyo.enigma.bean.*
 import homes.gensokyo.enigma.util.CiperTextUtil
 import homes.gensokyo.enigma.util.AppConstants
 import homes.gensokyo.enigma.util.DateUtils
+import homes.gensokyo.enigma.util.LogUtils
 import homes.gensokyo.enigma.util.SettingUtils.get
 import homes.gensokyo.enigma.util.SettingUtils.sharedPreferences
 import homes.gensokyo.enigma.util.TextUtils.toast
@@ -53,12 +55,13 @@ class UsrdataModel(repository1: UsrdataModelFactory, private val repository: Use
 
     //定时刷新任务
     private fun startPeriodicRefresh(headers: Map<String, String>) {
-        val intervalMillis: Long? = get("updateRate","15000").toLongOrNull()
+        val intervalMillis: Long? = get("updateRate","60000").toLongOrNull()
+        //TODO 这里实际上不是和Setting里面统一的
         viewModelScope.launch {
             flow {
                 while (true) {
                     emit(Unit)
-                    Log.d("startPeriodicRefresh", intervalMillis.toString())
+                    LogUtils.d("startPeriodicRefresh", intervalMillis.toString())
                     if (intervalMillis != null) {
                         delay(intervalMillis)
                     }
@@ -72,23 +75,26 @@ class UsrdataModel(repository1: UsrdataModelFactory, private val repository: Use
     private fun findStudentIndex(resultKid: List<Student>?): Int {
         val savedName = get("studentName","默认名字")
         resultKid?.forEachIndexed { index, student ->
-            //Log.d("findStudentIndex", "savedName: ${savedName}")
+            //LogUtils.d("findStudentIndex", "savedName: ${savedName}")
             if (student.studentName == savedName) {
-                Log.d("findStudentIndex", "savedName: ${savedName} + $index")
+                LogUtils.d("findStudentIndex", "savedName: ${savedName} + $index")
                 return index
             }
         }
 
         return -1
     }
-
+    val dashboardUpdateLimit = get("dashboardUpdateLimit", -50)
     suspend fun refreshData(headers: Map<String, String>) {
         viewModelScope.launch {
             try {
+                _studentData.postValue(
+                    DataState.Loading
+                )
                 val cipherText = CiperTextUtil.encrypt(get("wxOaOpenid","000"))
                 val resultGetRole = repository.fetchRole(cipherText, AppConstants.headerMap)
                 resultGetRole?.let {
-                    Log.d("UsrMdl", "Received Role info: $it ；$cipherText   11"  +get("wxOaOpenid","000").toString())
+                    LogUtils.d("UsrMdl", "Received Role info: $it ；$cipherText   11"  +get("wxOaOpenid","000").toString())
                 }
 
                 val resultLogin = repository.doLogin(AppConstants.headerMap)
@@ -99,27 +105,27 @@ class UsrdataModel(repository1: UsrdataModelFactory, private val repository: Use
                     .setCopyPersonCode(get("kidUuid","111"))
                     .setTypeCode(listOf(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11))
                     .build()
-                Log.d("queryData", "Query info: $qrBuild")
+                LogUtils.d("queryData", "Query info: $qrBuild")
 
                 val resultQuery = repository.queryData(AppConstants.headerMap, qrBuild )!!
                 resultQuery?.let {
-                    Log.d("queryData", "Received Query info: $it")
+                    LogUtils.d("queryData", "Received Query info: $it")
                 }
                 _queryData.postValue(resultQuery)
-                //Log.d("queryData", "Received Query info: $resultQuery")
+                //LogUtils.d("queryData", "Received Query info: $resultQuery")
 
                 val resultKid = repository.fetchStudents(AppConstants.headerMap)
 
                 resultKid?.joinToString(separator = "\n", prefix = "Students:\n") { student ->
                     "Name: ${student.studentName}, ID: ${student.studentId}, CN: ${student.cardNumber}"
-                }?.let { Log.d("StudentList", it) }
+                }?.let { LogUtils.d("StudentList", it) }
                 //这里智威后台发癫，会返回所有同一parent的kid，并且kid顺序有变化
 
-
+                LogUtils.d("UsrDataMdl", "$dashboardUpdateLimit")
                 /*
                 val resultBalance = repository.fetchBalance(AppConstants.headerMap)
                 if (resultBalance != null) {
-                    Log.d("UsrMdl", "Received Balance info: $resultBalance")
+                    LogUtils.d("UsrMdl", "Received Balance info: $resultBalance")
 
                 } else {
                     Log.e("UsrMdl", "Failed to fetch balance") // 记录错误信息
@@ -127,7 +133,7 @@ class UsrdataModel(repository1: UsrdataModelFactory, private val repository: Use
 
                 }
 
-                Log.d("startDat",  DateUtils.Date2Str(-10,true) )
+                LogUtils.d("startDat",  DateUtils.Date2Str(-10,true) )
                 val memberFlowRequest = MemberFlowJsonBuilder(
                     get("kidUuid","1111"),
                     listOf(2, 5, 6, 7),
@@ -160,7 +166,7 @@ class UsrdataModel(repository1: UsrdataModelFactory, private val repository: Use
                         7,
                         1,
                         1000,
-                        DateUtils.Date2Str(-360,true),
+                        DateUtils.Date2Str(dashboardUpdateLimit,false),
                         DateUtils.Date2Str(1)
                     )
                     repository.fetchMemberFlow(memberFlowAllRequest, AppConstants.headerMap)
@@ -178,30 +184,28 @@ class UsrdataModel(repository1: UsrdataModelFactory, private val repository: Use
                 val resultBalance = resultBalanceDeferred.await()!!
                 val resultMemberFlowAll = resultMemberFlowAllDeferred.await()
                 val resultMemberFlow = resultMemberFlowDeferred.await()
-                if(get("unilateralDeclarationCardNumber","fake") != resultBalance.cardNumber){
-                    //Log.d("UsrDataModel", get("unilateralDeclarationCardNumber","fake") + " != " + resultBalance.cardNumber)
-
-                    val editor = sharedPreferences!!.edit()
-                    editor.clear()
-                    editor.apply()
-                    "卡号不正确，强制退出！".toast()
-                    editor.clear()
-                    editor.commit()
-                    /*
-                    if (sharedPreferences.all.isEmpty()) {
-                        Log.d("Preferences", "清除成功")
-                    } else {
-                        Log.d("Preferences", "清除失败")
-                    }
-
-                     */
-
-                    restartApp()
+                if(!BuildConfig.DEBUG){
+                    if(get("unilateralDeclarationCardNumber","fake") != resultBalance.cardNumber) {
+                        //LogUtils.d("UsrDataModel", get("unilateralDeclarationCardNumber","fake") + " != " + resultBalance.cardNumber)
+                        val editor = sharedPreferences!!.edit()
+                        editor.clear()
+                        editor.apply()
+                        "卡号不正确，强制退出！".toast()
+                        editor.clear()
+                        editor.commit()
+                        /*
+                        if (sharedPreferences.all.isEmpty()) {
+                            LogUtils.d("Preferences", "清除成功")
+                        } else {
+                            LogUtils.d("Preferences", "清除失败")
+                        }
+                         */
+                        restartApp() }
                 }
 
 
 
-                Log.d("UsrDataModel", "Received MemberFlow info: $resultMemberFlow")
+                LogUtils.d("UsrDataModel", "Received MemberFlow info: $resultMemberFlow")
                 if (resultBalance != null && resultKid != null) {
                     val studentIndex = findStudentIndex(resultKid)
                     val studentName = resultKid[studentIndex].studentName ?: "默认姓名"
@@ -210,7 +214,7 @@ class UsrdataModel(repository1: UsrdataModelFactory, private val repository: Use
                     val headSculpture = resultKid[studentIndex].headSculpture ?: ""
 
                     if (resultMemberFlow != null) {
-                        Log.d("refreshData", resultMemberFlow.toString())
+                        LogUtils.d("refreshData", resultMemberFlow.toString())
                         _memberFlow.postValue(resultMemberFlow)
                         _memberFlowAll.postValue(resultMemberFlowAll)
                         _studentData.postValue(
@@ -233,7 +237,7 @@ class UsrdataModel(repository1: UsrdataModelFactory, private val repository: Use
                     _studentData.postValue(
                         DataState.Error("err")
                     )
-                    Log.d("refreshData", "resultBalance is null")
+                    LogUtils.d("refreshData", "resultBalance is null")
 
 
                 }
